Add `:legacy:di` module and move `DI`
