+++ legacy/common/src/main/java/com/fsck/k9/CommonApp.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import android.content.res.Configuration
 import android.content.res.Resources
 import app.k9mail.feature.widget.message.list.MessageListWidgetManager
+import app.k9mail.legacy.di.DI
 import com.fsck.k9.controller.MessagingController
 import com.fsck.k9.job.WorkManagerConfigurationProvider
 import com.fsck.k9.notification.NotificationChannelManager
