+++ legacy/core/src/main/java/com/fsck/k9/Core.kt
@@ -3,6 +3,8 @@ package com.fsck.k9
 import android.content.ComponentName
 import android.content.Context
 import android.content.pm.PackageManager
+import app.k9mail.legacy.di.EarlyInit
+import app.k9mail.legacy.di.inject
 import com.fsck.k9.job.K9JobManager
 import com.fsck.k9.mail.internet.BinaryTempFileBody
 import com.fsck.k9.notification.NotificationController
