+++ legacy/core/src/main/java/com/fsck/k9/mailstore/LocalStore.java
@@ -27,7 +27,7 @@ import android.text.TextUtils;
 
 import androidx.core.database.CursorKt;
 import com.fsck.k9.Account;
-import com.fsck.k9.DI;
+import app.k9mail.legacy.di.DI;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.controller.MessagingControllerCommands.PendingCommand;
 import com.fsck.k9.controller.PendingCommandSerializer;
