+++ legacy/core/src/main/java/com/fsck/k9/mailstore/MessageListCache.kt
@@ -1,6 +1,6 @@
 package com.fsck.k9.mailstore
 
-import com.fsck.k9.DI
+import app.k9mail.legacy.di.DI
 import com.fsck.k9.mail.Flag
 import kotlin.collections.set
