+++ legacy/core/src/main/java/com/fsck/k9/service/DatabaseUpgradeService.java
@@ -11,7 +11,7 @@ import android.os.IBinder;
 import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import com.fsck.k9.Account;
-import com.fsck.k9.DI;
+import app.k9mail.legacy.di.DI;
 import com.fsck.k9.K9;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.mail.power.PowerManager;
