+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/ChooseIdentity.java
@@ -10,7 +10,7 @@ import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.Toast;
 import com.fsck.k9.Account;
-import com.fsck.k9.DI;
+import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.account.Identity;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.ui.R;
