+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/compose/IdentityAdapter.java
@@ -12,7 +12,7 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 
 import com.fsck.k9.Account;
-import com.fsck.k9.DI;
+import app.k9mail.legacy.di.DI;
 import app.k9mail.legacy.account.Identity;
 import com.fsck.k9.Preferences;
 import com.fsck.k9.ui.R;
