+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/messagelist/MlfUtils.java
@@ -6,7 +6,7 @@ import java.util.List;
 import android.text.TextUtils;
 
 import com.fsck.k9.Account;
-import com.fsck.k9.DI;
+import app.k9mail.legacy.di.DI;
 import com.fsck.k9.controller.MessageReference;
 import com.fsck.k9.helper.Utility;
 import com.fsck.k9.mail.MessagingException;
